Fix compiler warnings for shadow and conversion (#52)
* fix: Enable and resolve compiler warnings

This commit enables the `-Wshadow`, `-Wconversion`, and `-Wsign-conversion` compiler flags and resolves all resulting warnings.

- Renamed shadowing variables in constructors and methods across multiple files (`command`, `target`, `state`, `serializer`, etc.) to avoid ambiguity, typically using an `initial_*` prefix.
- Made implicit narrowing and sign-changing conversions explicit by adding `static_cast` where appropriate. This improves code safety and clarity.
- Suppressed warnings originating from the third-party `fpm` fixed-point library by wrapping its include directives with `#pragma GCC diagnostic ignored`. This avoids modifying external code while maintaining a clean build.

* fix: Enable and resolve all compiler warnings

This commit enables the `-Wshadow`, `-Wconversion`, and `-Wsign-conversion` compiler flags and resolves all resulting warnings across the entire project.

- Renamed shadowing variables in constructors and methods across multiple files (`command`, `target`, `state`, `serializer`, `dualStepperDPTStartWithRadar.ino`, etc.) to avoid ambiguity, typically using an `initial_*` prefix.
- Made implicit narrowing and sign-changing conversions explicit by adding `static_cast` where appropriate. This improves code safety and clarity.
- Suppressed warnings originating from the third-party `fpm` fixed-point library by wrapping its include directives with `#pragma GCC diagnostic ignored`. This avoids modifying external code while maintaining a clean build.
- Updated the `IOWrapper` class to use `std::streamsize` for its `readsome` method to ensure type consistency with the underlying stream operations.

* fix: Enable and resolve all compiler warnings

This commit enables the `-Wshadow`, `-Wconversion`, and `-Wsign-conversion` compiler flags and resolves all resulting warnings across the entire project.

- Renamed shadowing variables in constructors and methods across multiple files (`command`, `target`, `state`, `serializer`, `dualStepperDPTStartWithRadar.ino`, etc.) to avoid ambiguity, typically using an `initial_*` prefix.
- Made implicit narrowing and sign-changing conversions explicit by adding `static_cast` where appropriate. This improves code safety and clarity.
- Suppressed warnings originating from the third-party `fpm` fixed-point library by wrapping its include directives in `tests/doctest_fpm_adapter.hpp` with `#pragma GCC diagnostic ignored`. This avoids modifying external code while maintaining a clean build.
- Updated the `IOWrapper` class to use `std::streamsize` for its `readsome` method to ensure type consistency with the underlying stream operations.

---------

+++ target.cpp
@@ -90,11 +90,11 @@ std::shared_ptr<const PositionVector> Target::interceptPosition() const {
 	// Quartic Coefficients
 	// Apply scaling shift to all coefficients uniformly
 	// This preserves the roots of the polynomial f(t) = 0
-	const localFixed c4 = localFixed::from_raw_value(c4_raw >> shift_amount);
-	const localFixed c3 = localFixed::from_raw_value(c3_raw >> shift_amount);
-	const localFixed c2 = localFixed::from_raw_value(c2_raw >> shift_amount);
-	const localFixed c1 = localFixed::from_raw_value(c1_raw >> shift_amount);
-	const localFixed c0 = localFixed::from_raw_value(c0_raw >> shift_amount);
+	const localFixed c4 = localFixed::from_raw_value(static_cast<int>(c4_raw >> shift_amount));
+	const localFixed c3 = localFixed::from_raw_value(static_cast<int>(c3_raw >> shift_amount));
+	const localFixed c2 = localFixed::from_raw_value(static_cast<int>(c2_raw >> shift_amount));
+	const localFixed c1 = localFixed::from_raw_value(static_cast<int>(c1_raw >> shift_amount));
+	const localFixed c0 = localFixed::from_raw_value(static_cast<int>(c0_raw >> shift_amount));
 
 	const std::function<localFixed(const localFixed&)> movingTargetInterceptQuartic =
 		[=](const localFixed& t1) -> localFixed {
@@ -118,7 +118,7 @@ std::shared_ptr<const PositionVector> Target::interceptPosition() const {
 	// L = -0.5*g
 	// So we want (diff.z + V_t.z*t + L*t^2)/t
 	const auto z_vel_num = diff.Z_coord + target_velocity.Z_coord * intercept_t -
-		localFixed::from_raw_value(L) * intercept_t * intercept_t;
+		localFixed::from_raw_value(static_cast<int>(L)) * intercept_t * intercept_t;
 
 	auto result = std::make_shared<PositionVector>(
 		x_vel_num / intercept_t,
